Move menu handlers into appManager
