+++ public/components/mainMenu.js
@@ -1,59 +1,17 @@
-const { dialog } = require('electron');
 const updater = require('./updater');
 
-const openDialogOptions = {
-  buttonLabel: 'Open',
-  nameFieldLabel: 'Open:',
-  defaultPath: 'Protocol.netcanvas',
-  filters: [{ name: 'Network Canvas', extensions: ['netcanvas'] }],
-  properties: ['openFile'],
-};
-
-const saveDialogOptions = {
-  buttonLabel: 'Save',
-  nameFieldLabel: 'Save:',
-  defaultPath: 'Protocol.netcanvas', // TODO: something based on existing filename?
-  filters: [{ name: 'Network Canvas', extensions: ['netcanvas'] }],
-};
-
-const openDialog = () =>
-  new Promise((resolve, reject) => {
-    dialog.showOpenDialog(openDialogOptions, (filename) => {
-      if (filename === undefined) { reject(); return; }
-      resolve(filename[0]);
-    });
-  });
-
-const openFile = appManager =>
-  () =>
-    openDialog()
-      .then(filePath => appManager.openFile(filePath));
-
-const saveDialog = () =>
-  new Promise((resolve, reject) => {
-    dialog.showSaveDialog(saveDialogOptions, (filename) => {
-      if (filename === undefined) { reject(); return; }
-      resolve(filename[0]);
-    });
-  });
-
-const saveCopy = appManager =>
-  () =>
-    saveDialog()
-      .then(filePath => appManager.saveCopy(filePath));
-
-const MenuTemplate = (appManager) => {
+const MenuTemplate = (handlers) => {
   const menu = [
     {
       label: 'File',
       submenu: [
         {
           label: 'Open...',
-          click: openFile(appManager),
+          click: handlers.openFile,
         },
         {
           label: 'Save copy...',
-          click: saveCopy(appManager),
+          click: handlers.saveCopy,
         },
       ],
     },
@@ -86,16 +44,7 @@ const MenuTemplate = (appManager) => {
         { type: 'separator' },
         {
           label: 'Clear storage data',
-          click: () => {
-            dialog.showMessageBox({
-              message: 'This will reset all app data, are you sure?',
-              buttons: ['OK', 'Cancel'],
-            }, (response) => {
-              if (response === 0) {
-                appManager.clearStorageData();
-              }
-            });
-          },
+          click: handlers.clearStorageData,
         },
       ],
     },
